requirements.txt
albumentations==2.0.5
datasets==3.3.2
matplotlib==3.10.1
numpy==1.26.4
Pillow==11.1.0
torch==2.6.0
torchvision==0.21.0
pipreqs
transformers==4.49.0
opencv-python
pytest